+++ test/support/firefox_profile_test.dart
@@ -19,7 +19,7 @@ library webdriver.support.firefox_profile_test;
 import 'dart:convert' show base64, Encoding, utf8;
 import 'dart:io' as io;
 
-import 'package:archive/archive.dart' show Archive, ArchiveFile, ZipDecoder;
+import 'package:archive/archive_io.dart' as archive;
 import 'package:test/test.dart';
 import 'package:webdriver/async_core.dart';
 import 'package:webdriver/support/firefox_profile.dart';
@@ -129,18 +129,19 @@ void main() {
       // ignore: deprecated_member_use_from_same_package
       profile.setOption(PrefsOption(Capabilities.hasNativeEvents, true));
 
-      final archive = unpackArchiveData(profile.toJson());
+      final zipArchive = unpackArchiveData(profile.toJson());
 
       final expectedFiles = ['prefs.js', 'user.js'];
-      expect(archive.files.length, greaterThanOrEqualTo(expectedFiles.length));
       expect(
-        archive.files,
-        anyElement((ArchiveFile f) => f.name == 'prefs.js'),
+          zipArchive.files.length, greaterThanOrEqualTo(expectedFiles.length));
+      expect(
+        zipArchive.files,
+        anyElement((archive.ArchiveFile f) => f.name == 'prefs.js'),
       );
 
       final prefs = FirefoxProfile.loadPrefsFile(MockFile(
         String.fromCharCodes(
-          archive.files.firstWhere((f) => f.name == 'user.js').content
+          zipArchive.files.firstWhere((f) => f.name == 'user.js').content
               as List<int>,
         ),
       ));
@@ -160,25 +161,25 @@ void main() {
       // ignore: deprecated_member_use_from_same_package
       profile.setOption(PrefsOption(Capabilities.hasNativeEvents, true));
 
-      final archive = unpackArchiveData(profile.toJson());
+      final zipArchive = unpackArchiveData(profile.toJson());
 
       final expectedFiles = [
         'prefs.js',
         'user.js',
         'addons.js',
-        'webapps/',
         'webapps/webapps.json'
       ];
-      expect(archive.files.length, greaterThanOrEqualTo(expectedFiles.length));
       expect(
-        archive.files,
-        anyElement((ArchiveFile f) => f.name == 'prefs.js'),
+          zipArchive.files.length, greaterThanOrEqualTo(expectedFiles.length));
+      expect(
+        zipArchive.files,
+        anyElement((archive.ArchiveFile f) => f.name == 'prefs.js'),
       );
 
       final prefs = FirefoxProfile.loadPrefsFile(
         MockFile(
           String.fromCharCodes(
-            archive.files.firstWhere((f) => f.name == 'user.js').content
+            zipArchive.files.firstWhere((f) => f.name == 'user.js').content
                 as List<int>,
           ),
         ),
@@ -195,9 +196,9 @@ void main() {
   }, timeout: const Timeout(Duration(minutes: 2)));
 }
 
-Archive unpackArchiveData(Map profileData) {
-  final zipArchive = base64.decode(profileData['firefox_profile'] as String);
-  return ZipDecoder().decodeBytes(zipArchive, verify: true);
+archive.Archive unpackArchiveData(Map profileData) {
+  final zipData = base64.decode(profileData['firefox_profile'] as String);
+  return archive.ZipDecoder().decodeBytes(zipData);
 }
 
 /// Simulate file for `FirefoxProfile.loadPrefsFile()`